Properly renamed most functions to SVG_

+++ src/common/bsp.cpp
@@ -1029,7 +1029,7 @@ static qboolean BSP_LoadPatchedPVS(bsp_t *bsp)
 	if (filebuf == 0)
 		return false;
 
-	size_t matrix_size = bsp->visrowsize * bsp->vis->numclusters;
+	size_t matrix_size = (size_t)bsp->visrowsize * (size_t)bsp->vis->numclusters;
 	if (filelen != matrix_size * 2)
 	{
 		FS_FreeFile(filebuf);
